requirements.txt
#setuptools==38.2.4
#zc.buildout==2.10.0
pytest~=6.2.5
flake8-todo
flake8-string-format
flake8-quotes
flake8-print
flake8-pep3101
flake8-isort
flake8-deprecated
flake8-debugger
flake8-commas
flake8-coding
flake8-blind-except
pymysql
colorlog
SQLAlchemy-Utils
netcdf4
pandas>=0.24
sphinxcontrib-serializinghtml
sphinxcontrib-qthelp
sphinx
zope.interfaces
zope.component

xarray~=0.19.0
numpy~=1.21.1
sqlalchemy~=1.4.25
addict~=2.4.0
scipy~=1.6.1
